Importacion/exportacion arreglada
Añadida la funcionalidad de la importacion y arreglada la exportacion

+++ src/jllmproyectadrian/model/JsonManager.java
@@ -6,11 +6,13 @@ package jllmproyectadrian.model;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
+import com.google.gson.stream.JsonReader;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.io.IOException;
+import java.io.StringReader;
 
 import java.nio.charset.StandardCharsets;
 import static jllmproyectadrian.util.CreateFolder.createFolderIfNotExists;
@@ -18,6 +20,10 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
+import java.text.SimpleDateFormat;
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 
 import java.util.logging.Level;
@@ -33,6 +39,16 @@ public class JsonManager {
     public JsonManager(DataBase dataB){
         
         this.dataBase = dataB;
+        Path folderPath = Paths.get(System.getProperty("user.home") + "\\Desktop\\JLLM\\IMPORTED_CONVERSATIONS");
+        if(createFolderIfNotExists(folderPath)){
+            System.out.println("Exito al crear carpeta");
+        }
+        
+        folderPath = Paths.get(System.getProperty("user.home") + "\\Desktop\\JLLM\\EXPORTED_CONVERSATIONS");
+        if(createFolderIfNotExists(folderPath)){
+            System.out.println("Exito al crear carpeta");
+        }
+        
         
     }
     
@@ -71,16 +87,11 @@ public class JsonManager {
         
     }
     
-    public ArrayList<Conversation> importTable(String tableName) throws FileNotFoundException, IOException{
+    public ArrayList<Conversation> importTable() throws FileNotFoundException, IOException{
         
         ArrayList<Conversation> conversations = new ArrayList<>();
         Path folderPath;
         folderPath = Paths.get(System.getProperty("user.home") + "\\Desktop\\JLLM\\IMPORTED_CONVERSATIONS");
-        if(createFolderIfNotExists(folderPath)){
-            System.out.println("Exito al crear carpeta");
-        }else{
-            System.out.println("Fallo al crear carpeta");
-        }
         Path jsonFilePath;
         File jsonFile = null;
         if (folderPath.toFile().exists() && folderPath.toFile().isDirectory()) {
@@ -99,38 +110,54 @@ public class JsonManager {
             System.out.println("La carpeta especificada no existe o no es un directorio válido.");
         }
         
+        if(jsonFile == null){
+            return null;
+        }
+        
         jsonFilePath = jsonFile.toPath();
-            
+        boolean failed = true;
         try (BufferedReader br = new BufferedReader(new FileReader(jsonFilePath.toFile()))) {
             String line;
             int row = 0;
-
             while ((line = br.readLine()) != null) {
                 row++;
                 try {
                     Conversation conver = new Conversation();
                     Gson gson = new Gson();
-                    JsonObject jsonObject = gson.fromJson(line, JsonObject.class);
+                    
+                    JsonReader reader = new JsonReader(new StringReader(line.toString()));
+                    reader.setLenient(true);
+                    
+                    JsonObject jsonObject = gson.fromJson(reader, JsonObject.class);
 
                     // Obtener valores específicos del JsonObject
                     String message = jsonObject.get("message").getAsString();
                     String answer = jsonObject.get("answer").getAsString();
-                    String date = jsonObject.get("date").getAsString();
-                    String time = jsonObject.get("time").getAsString();
+                    String dateString = jsonObject.get("date").getAsString();
+                    DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd-MM-yyyy");
+                    LocalDate date = LocalDate.parse(dateString, formatter);
+                    String timeString = jsonObject.get("time").getAsString();
+                    formatter = DateTimeFormatter.ofPattern("HH:mm:ss");
+                    LocalTime time = LocalTime.parse(timeString, formatter);
                     conver.setMessage(message);
                     conver.setAnswer(answer);
-                    conver.setDate(date, time);
+                    conver.setDateFromStrings(date.toString(), time.toString());
                     conversations.add(conver);
+                    failed = false;
                     
                 } catch (Exception ex) {
-                    //System.err.println("Error procesando línea " + row + ": " + ex.getMessage());
+                    System.err.println("Error procesando línea " + row + ": " + ex.getMessage());
+                    failed = true;
                 }
             }
         } catch (IOException e) {
             e.printStackTrace();
         }
-        System.out.println("Importacion realizada con exito");
-        jsonFile.delete();
+        if(!failed){
+            System.out.println("Importacion realizada con exito");
+            jsonFile.delete();
+        }
+        
         return conversations;
     }
